Directory navigation-modify serve_dir_with_tree()
((I'll bet someone who knows what they're doing can do this the right way instead of this way))
For use with tree:
Until this fork, directories could not be drilled down into because the URL was not managed.  This fork allows drilling down, but until someone devises a way to traverse back up, the browser's "back" button will have to be used.  The implication is that the usefulness is limited to sessions where the user began the session from the root-most directory they will want to navigate.  Other than that, it does provide a huge convenience for situations where multiple levels of directories need to be navigated.

Please, if someone can figure out how to navigate root-ward via a hyperlink rather than the back button, finish this concept.  THANK YOU!